requirements.txt
Flask
supabase
environs
google-generativeai
Pillow